add bean nams to redis config and qualifier to use of redisTemplate
