+++ src/main/java/uk/gov/ons/ctp/response/collection/exercise/CollectionExerciseApplication.java
@@ -24,6 +24,11 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.DependsOn;
 import org.springframework.context.annotation.Primary;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
+import org.springframework.data.redis.connection.jedis.JedisClientConfiguration;
+import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.integration.annotation.IntegrationComponentScan;
 import org.springframework.integration.annotation.MessagingGateway;
 import org.springframework.integration.annotation.ServiceActivator;
@@ -64,6 +69,32 @@ public class CollectionExerciseApplication {
   @Autowired private DataSource dataSource;
 
   @Autowired private StateTransitionManagerFactory collectionExerciseStateTransitionManagerFactory;
+
+  @Bean(name = "jedisConnectionFactory")
+  public JedisConnectionFactory jedisConnectionFactory() {
+    RedisStandaloneConfiguration configuration =
+        new RedisStandaloneConfiguration("localhost", 6379);
+    configuration.setDatabase(3);
+    JedisClientConfiguration jedisClientConfiguration =
+        JedisClientConfiguration.builder().usePooling().build();
+    JedisConnectionFactory factory =
+        new JedisConnectionFactory(configuration, jedisClientConfiguration);
+    factory.afterPropertiesSet();
+    return factory;
+  }
+
+  @Bean(name = "redisTemplate")
+  public RedisTemplate<String, Object> redisTemplate(
+      JedisConnectionFactory jedisConnectionFactory) {
+    final RedisTemplate<String, Object> redisTemplate = new RedisTemplate<>();
+    redisTemplate.setConnectionFactory(jedisConnectionFactory);
+    return redisTemplate;
+  }
+
+  @Bean
+  public RedisConnectionFactory redisConnectionFactory() {
+    return jedisConnectionFactory();
+  }
 
   /**
    * Bean used to map exceptions for endpoints
