+++ src/main/java/uk/gov/ons/ctp/response/collection/exercise/lib/common/redis/RedisUtil.java
@@ -1,12 +1,15 @@
 package uk.gov.ons.ctp.response.collection.exercise.lib.common.redis;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.core.RedisTemplate;
 
-@Configuration
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.stereotype.Component;
+
+@Component
 public class RedisUtil<Object> {
 
+  @Qualifier("redisTemplate")
   private RedisTemplate<String, String> redisTemplate;
 
   @Autowired
